requirements.txt
neo4j>=5.20.0
pyyaml>=6.0.1
tqdm>=4.66.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
numpy>=1.26.0
groq>=0.11.0